# fix: Incorrect duration and animate utilities (close #404)