### Description 📖 

This pull request fixes the problems described in #404.

Added a test that ensures these invalid usages do not generate CSS rules.

### Background 📜 

It seems that processing `<` tokens opened up the door for a new category of bugs in utilities that were created prior to the change supporting `<sm` and similar variants. A similar bug is #403.

Perhaps instead of only fixing it at the utility level on a case-by-case basis, these tokens should __not__ be interpreted as utilities in the first place:

```js
// extract.ts line 61
const matches = !className.startsWith('<') && className.match(/\w+/);
```

However, the `duration.value` case suggests that the regex in the plugin utils should be more strict, as `.` should only occur if there's also a `-` in the token, as in `ml-1.5`. Am I missing a case here?

### Test Case

#### Before

```css
  .\<transition {
    -webkit-transition-property: background-color, border-color, color, fill, stroke, opacity, -webkit-box-shadow, -webkit-transform, filter, backdrop-filter;
    -o-transition-property: background-color, border-color, color, fill, stroke, opacity, box-shadow, transform, filter, backdrop-filter;
    transition-property: background-color, border-color, color, fill, stroke, opacity, box-shadow, -webkit-box-shadow, transform, -webkit-transform, filter, backdrop-filter;
    -webkit-transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
    -o-transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
    transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
    -webkit-transition-duration: 150ms;
    -o-transition-duration: 150ms;
    transition-duration: 150ms;
  }
  .\.duration {
    -webkit-transition-duration: 150ms;
    -o-transition-duration: 150ms;
    transition-duration: 150ms;
  }
  .duration {
    -webkit-transition-duration: 150ms;
    -o-transition-duration: 150ms;
    transition-duration: 150ms;
  }
  .duration\.value {
    -webkit-transition-duration: 150ms;
    -o-transition-duration: 150ms;
    transition-duration: 150ms;
  }
  .\<animate {
    -webkit-animation-iteration-count: 1;
    animation-iteration-count: 1;
  }
```

#### After

No CSS rules