@antfu 2021-07-29
I think we might need a better solutions to this. Ideally, `.` and `<` is only used in `.dark` and `<sm`, for other utilities, we should not have them detected.