@ElMassimo 2021-07-29
Agreed. Feel free to modify this PR, the added test should be useful to prevent regressions.

Ideally, the plugin does not even send these invalid tokens to the compiler, but the compiler could also filter them out in a first pass in `extract.ts`.